# The minimal region cache