--- a/src/backoff.rs
+++ b/src/backoff.rs
@@ -5,7 +5,7 @@
 use rand::{thread_rng, Rng};
 use std::time::Duration;
 
-pub const DEFAULT_REGION_BACKOFF: Backoff = Backoff::no_jitter_backoff(2, 100, 30);
+pub const DEFAULT_REGION_BACKOFF: Backoff = Backoff::no_jitter_backoff(2, 500, 10);
 pub const OPTIMISTIC_BACKOFF: Backoff = Backoff::no_jitter_backoff(2, 500, 10);
 pub const PESSIMISTIC_BACKOFF: Backoff = Backoff::no_backoff();