--- a/src/pd/mod.rs
+++ b/src/pd/mod.rs
@@ -2,4 +2,4 @@ mod client;
 mod retry;
 
 pub use client::{PdClient, PdRpcClient};
-pub use retry::RetryClient;
+pub use retry::{RetryClient, RetryClientTrait};