The PR adds a minimal region cache implementation. It's far away from maturity. It's intended to discuss how the structure should look.

Related issue: #299 

## Main changes:
- Place the region cache as a component in `PdRpcClient`. `PdRpcClient` -> region cache -> retry client (the low level pd client)
- Make single-region plan a special case of multi-region plan

Credit to #116, #121 and #139. Since there have been some refactoring after the previous work, I didn't continue development based on those commits.

## Naive manual benchmark

Environment: single machine, 1 PD, 3 TiKV, default scheduler + shuffle-region + shuffle-leader
Workload: 6 threads, 512 workers,  50% single-key-read, 50% single-key-write

- zipfian distribution, pessimistic txns

master: TPS: 1943.9. Latency: mean: 41ms, P99: 1018ms, maximum: 1815ms
This PR: TPS 2108.5. Latency: mean: 12ms, P99: 108ms, maximum: 1061ms

- uniform distribution, optimistic txns

master: TPS: 4279.4, mean: 70ms, P99: 1048ms, maximum: 2601ms
This PR: TPS: 21129.8, mean: 22ms, P99: 44ms, maximum: 286ms

## Todo
- [x] Remove `read_through_cache` parameters, if we can always know the invalid cache entry
- [x] Treat single-region as a special case of multi-region
- [ ] Test
- [ ] Save unnecessary PD requests if they are querying the same region
- [ ] Finer grained error handling
- [ ] Naming