@ekexium 2021-07-01
> By the way, the backoff design is a bit from the client go. It is simpler because it always use the same backoff type (for different errors) and does not record the backoff history. These features are nice to have, but not an urgent task.

Yep, there is an issue there #310